requirements.txt
flask
pywebview
mutagen
sqlalchemy
